Remote Procedure Call System and UDP Virtual Connection improvements
RPCs are now implemented. It is checked if, after some specified time,
no acknowledge for their package was received, in which case the RPC
package is resent (with the same msg id but different seq id).

Virtual Connection was improved with unittests to verify the seq ID
calculus routines. The NetworkClient was implemented, completing
a networked client-server model.