Rewrote debugging output facility
In order to reduce code duplication, std::ostream is used through
a thin preprocessor interface. By this, the debug stream is run-time
configurable (through environment variables) and can be permanently
disabled at compile time.

The compiler optimizes out the code if permanently disabled. If disabled
in run-time, there is one if per debug statement, so it is rather cheap.