Implemented campaign::ScriptEngine
A very thin layer around the AngelScript library, mostly
for preparing the script engine similarily for the client
and server without code duplication.

It loads scripts sources via the ScriptFileManager and implements
the ScriptBuilder Include-Callback. By this, the game scripts can
include files from the campaign, even if the campaign is ZIPed.