Exposed API to create GameObject from AngelScript
Using the global gameObjectCreate function a GameObject, linked to a
certain AngelScript class can be created. Another function is used to
remove the Object from the ObjectManager, releasing the reference to the
AngelScript object. The object will not be rendered and noticed from the
(C++) application, the AS object will be destroyed as soon as all AS
references are released (of which shouldn't be any, anyway).