Implemented stub client/server negotiation
On connection, a client has to send some information about itself to
the server, which in turn decides to accept or refuse the connection.

For testing purposes, a rather dumb implementation is presented using
the RPC system. It will connect any client, version numbers and the
campaign are hard-coded. The client name will be stored, though, but
it is not checked for duplicates.

A "correct" implementation should implement the RPC handlers in the
campaign::ServerLogic and campaign::ClientLogic classes.