Changed game object protocol for AngelScript
Classes are registered as types, with information how to serialize them.
Using a function call, a new instance of an object can be created.

Preeviously, a class was instantiated and then coupled to a new object,
now it is the other way round.

Two helper functions for querying objects were implemented (findObject,
objectExists) to use the objects in AS.