Implemented join server acknowledge
When the server responds to a client's join request with his new
ID and the campaign to load. Then, the client checks if the campaign
is available, tries to load it and answers with an acknowledge RPC,
indicating whether he can join the server (e.g. map version imcompatibility,
map not installed etc).