"Fixed" AngelScript bug
The AngelScript header file is agnostic to clang, so the support macro
was set manualy in the main CMakeLists. But this caused another error
to be thrown, idk if it will fail on gcc/msvc or any other compiler. It
does work with clang, so I don't care until someone starts screaming!