Added server and client dependent object scripts
Formerly, script objects of game objects used the same classes (i.e. class name)
on server and clients, which is might be a bad thing. Now, it is possible
to have different class names. The object manager is told if it manages
 a server or a client, which might come in handy for future improvements.

It is still possible to use the same class name for client and server,
but you should make sure that they have different "implementations", e.g.
through the preprocessor of AngelScript.