Added paging to editor
Terrains are only loaded if they are close to the camera. If a modified
terrain needs to be unloaded, it is stored in a temporary file. Saving
was also implemented for testing purposes. Loaded terrains to be saved
are just written to disk, unloaded terrains are copied over from the
temporary terrain file.

A nice progress bar in the windows status bar shows the saving progress.
Saving is delayed until all updates are calculated, indicated by a
"scrolling progress bar", editing is disabled during that time.