Define terrain tiles based on slot definition
CampaignManager stores a list of slots where we want a
terrain tile. In the TerrainDefiner class, we now try to
create a height-zero terrain if we do not find a tmp terrain or
a normal terrain file.

The next step is exporting a mechanism to edit the required slots,
so we can specify the shape and extents of our terrain.